Moved ClickType to its own file
