+++ src/com/projectkorra/ProjectKorra/Utilities/ClickType.java
@@ -1,0 +1,5 @@
+package com.projectkorra.ProjectKorra.Utilities;
+
+public enum ClickType {
+	SHIFTDOWN, SHIFTUP, LEFTCLICK, RIGHTCLICK
+}
